// server/src/types/errors/errorCodes.ts
export enum DefaultErrorCodes {
  API_ERROR = 'API_ERROR',
  APOLLO_API_ERROR = 'APOLLO_API_ERROR',
  AUTH_REQUIRED = 'AUTH_REQUIRED',
  ACCESS_DENIED = 'ACCESS_DENIED',
  BAD_REQUEST = 'BAD_REQUEST',
  UNAUTHORIZED = 'UNAUTHORIZED',
  RESOURCE_NOT_FOUND = 'RESOURCE_NOT_FOUND',
  RESOURCE_HAS_CONFLICT = 'RESOURCE_HAS_CONFLICT',
  UNPROCESSABLE_ENTITY = 'UNPROCESSABLE_ENTITY',
  RESOURCE_LOCKED = 'RESOURCE_LOCKED',
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
  GONE = 'GONE',
  NOT_MODIFIED = 'NOT_MODIFIED',
  TOO_MANY_REQUESTS = 'TOO_MANY_REQUESTS',
  PAYLOAD_TOO_LARGE = 'PAYLOAD_TOO_LARGE',
}